requirements.bert.txt
bert-serving-client==1.10.0